refactor(ui): make the connection create form embeddable
- split the 650-line template-create-form into a thin dialog wrapper, a
  layout-agnostic form body, a create/authorize submit hook, and a pure
  payload builder
- extract LabeledInput, OAuthAppHint, OverridableSection and field copy
  tables into their own files
- add a shared compact() util for dropping undefined payload fields
