+++ packages/ui/src/modules/connections/lib/connection-name.ts
@@ -1,0 +1,16 @@
+import { connectionNameSchema } from "api-server-api";
+
+export function validateConnectionName(name: string): string | null {
+  const result = connectionNameSchema.safeParse(name);
+  return result.success
+    ? null
+    : (result.error.issues[0]?.message ?? "Invalid name");
+}
+
+export function slugifyTemplateName(name: string): string {
+  return name
+    .toLowerCase()
+    .replace(/[^a-z0-9]+/g, "-")
+    .replace(/^-+|-+$/g, "")
+    .slice(0, 63);
+}
